Adjusted to Bang to apply single responsibility on all the functions. Removed all unused imports. moved BangCache.kt and BangResult.kt into their respective files in Service.
