+++ src/main/kotlin/com/wink/dbse/DBSEApplication.kt
@@ -3,7 +3,6 @@ package com.wink.dbse
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.boot.runApplication
-import org.springframework.scheduling.annotation.EnableAsync
 import org.springframework.scheduling.annotation.EnableScheduling
 
 @SpringBootApplication
