+++ src/main/kotlin/com/wink/dbse/command/admin/Purge.kt
@@ -2,7 +2,6 @@ package com.wink.dbse.command.admin
 
 import com.jagrosh.jdautilities.command.Command
 import com.jagrosh.jdautilities.command.CommandEvent
-import com.wink.dbse.eventlistener.ChannelOptionReactionAdder
 import net.dv8tion.jda.api.Permission
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
