+++ src/main/kotlin/com/wink/dbse/command/economy/Profile.kt
@@ -2,9 +2,9 @@ package com.wink.dbse.command.economy
 
 import com.jagrosh.jdautilities.command.Command
 import com.jagrosh.jdautilities.command.CommandEvent
-import com.wink.dbse.command.game.BangResult
 import com.wink.dbse.entity.BangEntity
 import com.wink.dbse.entity.UserEntity
+import com.wink.dbse.entity.game.BangResult
 import com.wink.dbse.extension.safeName
 import com.wink.dbse.repository.BangRepository
 import com.wink.dbse.repository.UserRepository
@@ -19,11 +19,11 @@ class Profile(
     private val messenger: Messenger,
     private val userRepository: UserRepository,
     private val bangRepository: BangRepository
-    ) : Command() {
+) : Command() {
 
     init {
         name = "profile"
-        help = "todo"
+        help = "shows your profile information"
     }
 
     private val eb = EmbedBuilder()
@@ -38,14 +38,14 @@ class Profile(
 
     private fun setUserInfo(user: UserEntity) {
         eb.addField("Streak", "${user.bangStreak}" + if(user.bangStreak > 0)" :fire:" else "", true)
-        eb.addField("Karma","${user.upVotes - user.downVotes}",true)
+        eb.addField("Karma","${user.upVotes - user.downVotes}",false)
         eb.addField("Wallet","${user.wallet} gc",true)
     }
 
     private fun generateBangInfo(bangs: List<BangEntity>) {
         var jams = 0
         var deaths = 0
-
+        //TODO: remove for the new database option
         for (bang in  bangs) {
             when (bang.result) {
                 BangResult.DIE.value-> deaths++
@@ -55,27 +55,21 @@ class Profile(
         setBangInfo(bangs.size, deaths, jams)
     }
 
-
     private fun setBangInfo (bangs: Int, deaths: Int, jams: Int) {
         eb.addField("Bangs", "$bangs",true)
         eb.addField("Deaths", "$deaths",true)
         eb.addField("Jams", "$jams",true)
-        eb.addField("Survival Rate",  "${getSurvivalRate(deaths.toFloat(), bangs)}%",true)
+        eb.addField("Survival Rate",  "${getSurvivalRate(deaths, bangs)}%",true)
     }
 
-    private fun getSurvivalRate (deaths: Float, bangs: Int): Double {
-        return if(bangs > 0) (100 - (deaths.toDouble().div(bangs) * 100 * 100).roundToInt() / 100.00) else 100.00
+    private fun getSurvivalRate (deaths: Int, bangs: Int): Double {
+        return if(bangs > 0) (100 - (deaths.toFloat().div(bangs) * 100 * 100).roundToInt() / 100.00) else 100.00
     }
 
     private fun setHeading (event: CommandEvent) {
         eb.setTitle( "${event.author.safeName()}'s Profile")
         eb.setThumbnail(event.author.avatarUrl)
-        eb.setColor(if(event.member != null) event.member.color else Color.LIGHT_GRAY)
+        eb.setColor(event.member?.color ?: Color.LIGHT_GRAY)
     }
-
-
-
-
-
 
 }
