+++ src/main/kotlin/com/wink/dbse/entity/BangEntity.kt
@@ -1,7 +1,10 @@
 package com.wink.dbse.entity
 
 import java.time.LocalDateTime
-import javax.persistence.*
+import javax.persistence.Entity
+import javax.persistence.GeneratedValue
+import javax.persistence.Id
+import javax.persistence.Table
 
 @Entity
 @Table(name = "bang")
