+++ src/main/kotlin/com/wink/dbse/entity/MessageEntity.kt
@@ -1,6 +1,9 @@
 package com.wink.dbse.entity
 
-import javax.persistence.*
+import javax.persistence.Column
+import javax.persistence.Entity
+import javax.persistence.Id
+import javax.persistence.Table
 
 @Entity
 @Table(name = "message")
