+++ src/main/kotlin/com/wink/dbse/entity/UserEntity.kt
@@ -14,5 +14,9 @@ class UserEntity(
         val wallet: Long = 0,
         val purchasedRoleId: Long? = null,
         val roleExpiry: LocalDateTime? = null,
-        val bangStreak: Int = 0
+        val bangStreak: Int = 0//,
+//        val bangAttempts: Int = 0,
+//        val bangDeaths: Int = 0,
+//        val bangJams: Int = 0
 )
+//TODO: add an annotation convert from one to many then compile to one value
