+++ src/main/kotlin/com/wink/dbse/eventlistener/MessageBulkDeleteLogger.kt
@@ -12,7 +12,6 @@ import net.dv8tion.jda.api.hooks.ListenerAdapter
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Component
-import java.lang.StringBuilder
 
 @Component
 class MessageBulkDeleteLogger(
