+++ src/main/kotlin/com/wink/dbse/eventlistener/MessageCollector.kt
@@ -2,11 +2,11 @@ package com.wink.dbse.eventlistener
 
 import com.jagrosh.jdautilities.command.CommandClient
 import com.wink.dbse.entity.MessageEntity
+import com.wink.dbse.repository.MessageRepository
 import com.wink.dbse.service.MessageConvertingService
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent
 import net.dv8tion.jda.api.hooks.ListenerAdapter
 import org.springframework.stereotype.Component
-import com.wink.dbse.repository.MessageRepository
 import javax.transaction.Transactional
 
 @Component
