+++ src/main/kotlin/com/wink/dbse/eventlistener/MessageDeleteLogger.kt
@@ -5,7 +5,6 @@ import com.wink.dbse.property.ChannelIds
 import com.wink.dbse.repository.MessageRepository
 import com.wink.dbse.service.LoggedMessageFormatter
 import com.wink.dbse.service.Messenger
-import net.dv8tion.jda.api.entities.MessageChannel
 import net.dv8tion.jda.api.entities.TextChannel
 import net.dv8tion.jda.api.entities.User
 import net.dv8tion.jda.api.events.message.MessageDeleteEvent
