+++ src/main/kotlin/com/wink/dbse/eventlistener/MessageUpdateLogger.kt
@@ -4,8 +4,8 @@ import com.wink.dbse.entity.MessageEntity
 import com.wink.dbse.property.ChannelIds
 import com.wink.dbse.repository.MessageRepository
 import com.wink.dbse.service.LoggedMessageFormatter
-import com.wink.dbse.service.Messenger
 import com.wink.dbse.service.MessageConvertingService
+import com.wink.dbse.service.Messenger
 import net.dv8tion.jda.api.entities.TextChannel
 import net.dv8tion.jda.api.entities.User
 import net.dv8tion.jda.api.events.message.MessageUpdateEvent
