+++ src/main/kotlin/com/wink/dbse/factory/CommandClientFactory.kt
@@ -6,7 +6,6 @@ import com.jagrosh.jdautilities.command.CommandClientBuilder
 import com.wink.dbse.property.BotProperties
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Bean
-import org.springframework.stereotype.Component
 import org.springframework.stereotype.Service
 
 @Service
