+++ src/main/kotlin/com/wink/dbse/factory/JDAFactory.kt
@@ -9,7 +9,6 @@ import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.context.annotation.Bean
 import org.springframework.stereotype.Service
-import java.lang.Exception
 import kotlin.system.exitProcess
 
 @Service
