+++ src/main/kotlin/com/wink/dbse/service/EconomyService.kt
@@ -8,7 +8,6 @@ import net.dv8tion.jda.api.entities.Guild
 import net.dv8tion.jda.api.entities.Role
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
-import java.lang.IllegalStateException
 import java.time.LocalDateTime
 
 @Service
